refactor: move injection tokens to shared module and update imports
