+++ packages/frameworks/vue/src/chat/index.ts
@@ -1,6 +1,5 @@
 export { default as GenuiChat } from './GenuiChat.vue';
 export * from './chat.types.js';
-export * from './injection-tokens.js';
 export * from './i18n/index.js';
 export * from './tiny-robot-patch/index.js';
 export * from './event-emitter.js';
